+++ NEWS.md
@@ -2,6 +2,8 @@
 
 * Added non-interactive versions of the addins, thanks to the V8 package.
 * The number of spaces for the indentation is now based on the RStudio preference.
+* Added a XML/SVG prettifier.
+* Added a C/C++/Java prettifier, requiring `clang-format`.
 
 
 # prettifyAddins 0.1.0